+++ src/Field/RadioList.php
@@ -34,6 +34,34 @@ final class RadioList extends PartsField implements ValidationClassInterface
         $this->widget = RadioListWidget::create('');
     }
 
+    public function radioWrapTag(?string $name): self
+    {
+        $new = clone $this;
+        $new->widget = $this->widget->radioWrapTag($name);
+        return $new;
+    }
+
+    public function radioWrapAttributes(array $attributes): self
+    {
+        $new = clone $this;
+        $new->widget = $this->widget->radioWrapAttributes($attributes);
+        return $new;
+    }
+
+    public function radioWrapClass(?string ...$class): self
+    {
+        $new = clone $this;
+        $new->widget = $this->widget->radioWrapClass(...$class);
+        return $new;
+    }
+
+    public function addRadioWrapClass(?string ...$class): self
+    {
+        $new = clone $this;
+        $new->widget = $this->widget->addRadioWrapClass(...$class);
+        return $new;
+    }
+
     public function radioAttributes(array $attributes): self
     {
         $new = clone $this;
@@ -59,6 +87,13 @@ final class RadioList extends PartsField implements ValidationClassInterface
     {
         $new = clone $this;
         $new->widget = $this->widget->addRadioLabelAttributes($attributes);
+        return $new;
+    }
+
+    public function radioLabelWrap(bool $wrap): self
+    {
+        $new = clone $this;
+        $new->widget = $this->widget->radioLabelWrap($wrap);
         return $new;
     }